project/django_playground/requirements.txt
celery==5.3.6
django==5.0.1
djangorestframework==3.14.0
django-debug-toolbar==4.3.0
redis==5.0.1